UPDATE: modificando a estrutura dos relatórios
