+++ src/main/java/br/com/systemsgs/ordem_servico_backend/config/GlobalApplicationConfiguration.java
@@ -12,6 +12,8 @@ public class GlobalApplicationConfiguration {
 
     public static final String[] ROTAS_PUBLICAS = {
             "/api/usuarios/v1/salvar",
+            "/api/clientes/v1/relatorio/excel",
+            "/api/clientes/v1/relatorio/pdf",
             "/api/usuarios/v1/login",
             "/api/health/check",
             "/swagger-ui/**",
