+++ src/main/java/br/com/systemsgs/ordem_servico_backend/relatorios/GerarRelatorio.java
@@ -1,11 +1,13 @@
-package br.com.systemsgs.ordem_servico_backend.relatorios.excel;
+package br.com.systemsgs.ordem_servico_backend.relatorios;
 
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.ResponseEntity;
 
 import java.io.IOException;
 
-public interface GerarRelatorioExcel {
+public interface GerarRelatorio {
 
     ResponseEntity<byte[]> gerarRelatorioExcel (HttpServletResponse response) throws IOException;
+
+    byte[] gerarRelatorioPdf() throws IOException;
 }
