--- src/main/java/br/com/systemsgs/ordem_servico_backend/relatorios/pdf/GerarRelatorioPdf.java
@@ -1,9 +1,0 @@
-package br.com.systemsgs.ordem_servico_backend.relatorios.pdf;
-
-import java.io.IOException;
-
-public interface GerarRelatorioPdf {
-
-    byte[] gerarRelatorioPdf() throws IOException;
-
-}
